4.5: Minor issue with deleting old thumbs when folder did not exist yet.

+++ src/nl/asymmetrics/droidshows/DroidShows.java
@@ -340,8 +340,10 @@ public class DroidShows extends ListActivity
 								try {
 									backupRestore(source, destination);
 									updateDS.updateDroidShows();
-									for (File thumb : new File(getApplicationContext().getFilesDir().getAbsolutePath() +"/thumbs/banners/posters").listFiles())
-									    thumb.delete();
+									File thumbs[] = new File(getApplicationContext().getFilesDir().getAbsolutePath() +"/thumbs/banners/posters").listFiles();
+									if (thumbs != null)
+										for (File thumb : thumbs)
+											thumb.delete();
 									for (File file : new File(getApplicationInfo().dataDir +"/databases/").listFiles())
 									    if (!file.getName().equalsIgnoreCase("DroidShows.db")) file.delete();
 									getSeries();
